Cheese-off Initial Commit
This is the initial commit for the Cheese-off, including the beginning
of the design.R file. This will be where the design is generated,
ideally functionalized.

Then, the functionalized files will be used to create a Shiny app that
can generate, save, and load design files for the final bracket.